suggest for expression command migration
