+++ src/platform/packages/shared/kbn-esql-language/src/commands/definitions/utils/index.ts
@@ -9,8 +9,6 @@
 
 export {
   getFieldsSuggestions,
-  getFunctionsSuggestions,
-  getLiteralsSuggestions,
   getControlSuggestionIfSupported,
   getControlSuggestion,
   getSafeInsertText,
